Move exceptions to separate package
