+++ src/main/java/uk/gov/hmcts/reform/jobscheduler/services/jobs/exceptions/JobActionSerializationException.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.jobscheduler.services.jobs.exceptions;
+
+public class JobActionSerializationException extends RuntimeException {
+
+    public JobActionSerializationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
