+++ src/main/java/uk/gov/hmcts/reform/jobscheduler/services/jobs/exceptions/JobException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.jobscheduler.services.jobs;
+package uk.gov.hmcts.reform.jobscheduler.services.jobs.exceptions;
 
 public class JobException extends RuntimeException {
