+++ src/test/java/uk/gov/hmcts/reform/jobscheduler/services/jobs/jobsservice/CreateJobTest.java
@@ -13,7 +13,7 @@ import org.quartz.Trigger;
 import uk.gov.hmcts.reform.jobscheduler.jobs.HttpCallJob;
 import uk.gov.hmcts.reform.jobscheduler.model.Job;
 import uk.gov.hmcts.reform.jobscheduler.services.jobs.ActionSerializer;
-import uk.gov.hmcts.reform.jobscheduler.services.jobs.JobException;
+import uk.gov.hmcts.reform.jobscheduler.services.jobs.exceptions.JobException;
 import uk.gov.hmcts.reform.jobscheduler.services.jobs.JobsService;
 
 import java.sql.Date;
